Basic calc logic achieved
Basic calc logic achieved I think.
Managed to have the calculator not evaluate more than a single pair of numbers at a time.

Managed to get the correct result from TOPs example equation to getting 42.
12 + 7 - 5 * 3 = should yield 42

Began playing with some CSS styling.

Next is to continue working on the 'Gotchas' outlined in the curriculum and the 'Extra Credit' implementations.